Refactored for better segregation of concerns
