+++ SkillMatrix/Controllers/AuthController.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using SkillMatrix.Application.DTOs.Identity;
 using SkillMatrix.Application.Extensions;
+using SkillMatrix.Application.Services.Authentication;
 using SkillMatrix.Domain.Users.Models;
 using System;
 using System.Collections.Generic;
@@ -23,17 +24,18 @@ namespace SkillMatrix.Application.Controllers
     public class AuthController : MainController
     {
         private readonly SignInManager<User> _signInManager;
-        private readonly UserManager<User> _userManager;
-        private readonly TokenSettings _tokenSettings;
+        private readonly IAuthenticationService _authenticationService;
+        private readonly ITokenService _tokenService;
+
 
         public AuthController(
-             SignInManager<User> signInManager,
-             UserManager<User> userManager,
-             IOptions<TokenSettings> tokenSettings)
+            SignInManager<User> signInManager,
+            IAuthenticationService authenticationService,
+            ITokenService tokenService)
         {
             _signInManager = signInManager;
-            _userManager = userManager;
-            _tokenSettings = tokenSettings.Value;
+            _authenticationService = authenticationService;
+            _tokenService = tokenService;
         }
 
         [HttpPost("new-account")]
@@ -41,18 +43,11 @@ namespace SkillMatrix.Application.Controllers
         {
             if (!ModelState.IsValid) return CustomResponse(ModelState);
 
-            var user = new User
-            {
-                UserName = registerUser.Email,
-                Email = registerUser.Email,
-                EmailConfirmed = true
-            };
-
-            var result = await _userManager.CreateAsync(user, registerUser.Password);
+            var result = await _authenticationService.CreateNewUser(registerUser);
 
             if (result.Succeeded)
             {
-                return CustomResponse(await GenerateJWT(registerUser.Email));
+                return CustomResponse(await _tokenService.GenerateJWT(registerUser.Email));
             }
 
             foreach (var error in result.Errors)
@@ -78,7 +73,7 @@ namespace SkillMatrix.Application.Controllers
             //lockoutOnFailure: lock the loging in the application for an amount of time if there were too many failed logins attempts. 
             if (result.Succeeded)
             {
-                return CustomResponse(await GenerateJWT(loginUser.Email));
+                return CustomResponse(await _tokenService.GenerateJWT(loginUser.Email));
             }
 
             if (result.IsLockedOut)
@@ -91,81 +86,5 @@ namespace SkillMatrix.Application.Controllers
 
             return CustomResponse();
         }
-
-        private async Task<UserLoginResponse> GenerateJWT(string email)
-        {
-            var user = await _userManager.FindByEmailAsync(email);
-            var claims = await _userManager.GetClaimsAsync(user);
-            var userRoles = await _userManager.GetRolesAsync(user);
-
-            var identityClaims = GenerateIdentityClaims(user, claims, userRoles);
-
-            var encodedToken = GenerateToken(identityClaims);
-
-            var response = new UserLoginResponse
-            {
-                AccessToken = encodedToken,
-                ExpiresIn = TimeSpan.FromHours(_tokenSettings.ExpirationHours).TotalSeconds,
-                UserToken = UserToken.Create(user, claims)
-            };
-
-            return response;
-        }
-
-        private string GenerateToken(ClaimsIdentity identityClaims)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_tokenSettings.Secret);
-
-            var token = tokenHandler.CreateToken(new SecurityTokenDescriptor
-            {
-                Issuer = _tokenSettings.Issuer,
-                Audience = _tokenSettings.ValidIn,
-                Subject = identityClaims,
-                Expires = DateTime.UtcNow.AddHours(_tokenSettings.ExpirationHours),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), algorithm: SecurityAlgorithms.HmacSha256Signature)
-            });
-
-            var encodedToken = tokenHandler.WriteToken(token);
-            return encodedToken;
-        }
-
-        private static ClaimsIdentity GenerateIdentityClaims(User user, IList<Claim> claims, IList<string> userRoles)
-        {
-            AddUserInfoToClaims(user, claims);
-            AddTokenInfoToClaims(claims);
-            AddUserRolesToClaims(claims, userRoles);
-
-            var identityClaims = new ClaimsIdentity();
-            identityClaims.AddClaims(claims);
-            return identityClaims;
-        }
-
-        private static void AddUserRolesToClaims(IList<Claim> claims, IList<string> userRoles)
-        {
-            foreach (var userRole in userRoles)
-            {
-                claims.Add(new Claim(type: "role", value: userRole));
-            }
-        }
-
-        private static void AddTokenInfoToClaims(IList<Claim> claims)
-        {
-            claims.Add(new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()));
-            claims.Add(new Claim(JwtRegisteredClaimNames.Nbf, ToUnixEpochDate(DateTime.UtcNow).ToString())); //Token expiration
-            claims.Add(new Claim(JwtRegisteredClaimNames.Iat, ToUnixEpochDate(DateTime.UtcNow).ToString(), ClaimValueTypes.Integer64)); //Token issue date
-        }
-
-        private static void AddUserInfoToClaims(User user, IList<Claim> claims)
-        {
-            claims.Add(new Claim(JwtRegisteredClaimNames.Sub, user.Id));
-            claims.Add(new Claim(JwtRegisteredClaimNames.Email, user.Email));
-        }
-
-        private static long ToUnixEpochDate(DateTime date)
-            => (long)Math.Round((date.ToUniversalTime() - new DateTimeOffset(1970, 1, 1, 0, 0, 0, TimeSpan.Zero)).TotalSeconds);
-
-
-
     }
 }
