+++ SkillMatrix/Services/Authentication/AuthenticationRegistry.cs
@@ -1,0 +1,41 @@
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
+using SkillMatrix.Application.DTOs.Identity;
+using SkillMatrix.Application.Extensions;
+using System.Text;
+
+namespace SkillMatrix.Application.Services.Authentication
+{
+    public static class AuthenticationRegistry
+    {
+        public static void ConfigureAuthentication(this IServiceCollection services, IConfiguration configuration)
+        {
+            var tokenSettings = configuration.GetSection("TokenSettings");
+            services.Configure<TokenSettings>(tokenSettings);
+
+            var token = tokenSettings.Get<TokenSettings>();
+            var key = Encoding.ASCII.GetBytes(token.Secret);
+
+            services.AddAuthentication(x =>
+            {
+                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+            }).AddJwtBearer(x =>
+            {
+                x.RequireHttpsMetadata = true;
+                x.SaveToken = true;
+                x.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(key),
+                    ValidateIssuer = true,
+                    ValidateAudience = true,
+                    ValidAudience = token.ValidAt,
+                    ValidIssuer = token.Issuer,
+                };
+            });
+        }
+    }
+}
