+++ SkillMatrix/Services/Authentication/TokenSettings.cs
@@ -10,6 +10,6 @@ namespace SkillMatrix.Application.Extensions
         public string Secret { get; set; }
         public int ExpirationHours { get; set; }
         public string Issuer { get; set; }
-        public string ValidIn { get; set; }
+        public string ValidAt { get; set; }
     }
 }
